“Hacks” to allow open text input to be correct and show model answer
These changes may be covering a bug in the ButtonsView object... If the
question is correct then it's not possible to view the model answer. The
onActionClicked ensures that the button is always enabled.

It also sets the value to correct if it can be submitted, i.e. if there's
any text in the textbox it's submittable and therefore correct.